+++ trixnity-client/trixnity-client-media-indexeddb/src/jsMain/kotlin/net/folivo/trixnity/client/media/indexeddb/IndexedDBMediaStore.kt
@@ -1,17 +1,29 @@
+@file:OptIn(ExperimentalWasmJsInterop::class)
+
 package net.folivo.trixnity.client.media.indexeddb
 
-import com.juul.indexeddb.Database
-import com.juul.indexeddb.Key
-import com.juul.indexeddb.openDatabase
 import js.buffer.ArrayBuffer
 import js.typedarrays.Uint8Array
-import kotlinx.coroutines.*
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.DelicateCoroutinesApi
+import kotlinx.coroutines.GlobalScope
+import kotlinx.coroutines.coroutineScope
 import kotlinx.coroutines.flow.FlowCollector
+import kotlinx.coroutines.job
+import kotlinx.coroutines.launch
+import kotlinx.coroutines.promise
 import net.folivo.trixnity.client.MatrixClientConfiguration
 import net.folivo.trixnity.client.MediaStoreModule
 import net.folivo.trixnity.client.media.CachedMediaStore
 import net.folivo.trixnity.client.media.MediaStore
-import net.folivo.trixnity.utils.*
+import net.folivo.trixnity.idb.utils.IDBUtils
+import net.folivo.trixnity.idb.utils.readTransaction
+import net.folivo.trixnity.idb.utils.writeTransaction
+import net.folivo.trixnity.utils.ByteArrayFlow
+import net.folivo.trixnity.utils.byteArrayFlowFromReadableStream
+import net.folivo.trixnity.utils.nextString
+import net.folivo.trixnity.utils.toByteArray
+import net.folivo.trixnity.utils.writeTo
 import org.koin.dsl.module
 import web.blob.Blob
 import web.events.EventType
@@ -19,6 +31,8 @@ import web.events.addEventHandler
 import web.http.BodyInit
 import web.http.Response
 import web.http.blob
+import web.idb.IDBDatabase
+import web.idb.IDBValidKey
 import web.streams.TransformStream
 import web.window.window
 import kotlin.random.Random
@@ -35,9 +49,9 @@ internal class IndexedDBMediaStore(
         const val TMP_MEDIA_OBJECT_STORE_NAME = "tmp"
     }
 
-    private lateinit var database: Database
+    private lateinit var database: IDBDatabase
     override suspend fun init(coroutineScope: CoroutineScope) {
-        database = openDatabase(databaseName, 2) { database, oldVersion, _ ->
+        database = IDBUtils.openDatabase(databaseName, 2) { database, oldVersion, _ ->
             if (oldVersion < 1) {
                 database.createObjectStore(MEDIA_OBJECT_STORE_NAME)
             }
@@ -77,30 +91,30 @@ internal class IndexedDBMediaStore(
         val value = Response(BodyInit(transformStream.readable)).blob()
         database.writeTransaction(MEDIA_OBJECT_STORE_NAME) {
             val store = objectStore(MEDIA_OBJECT_STORE_NAME)
-            store.put(value, Key(url))
+            store.put(value, IDBValidKey(url))
         }
     }
 
     override suspend fun getMedia(url: String): IndexeddbPlatformMedia? =
-        database.transaction(MEDIA_OBJECT_STORE_NAME) {
+        database.readTransaction(MEDIA_OBJECT_STORE_NAME) {
             val store = objectStore(MEDIA_OBJECT_STORE_NAME)
-            store.get(Key(url))
-        }.unsafeCast<Blob?>()
+            store.get<Blob>(IDBValidKey(url))
+        }
             ?.let { BlobBasedIndexeddbPlatformMediaImpl(url, it) }
 
     override suspend fun deleteMedia(url: String): Unit =
         database.writeTransaction(MEDIA_OBJECT_STORE_NAME) {
             val store = objectStore(MEDIA_OBJECT_STORE_NAME)
-            store.delete(Key(url))
+            store.delete(IDBValidKey(url))
         }
 
     override suspend fun changeMediaUrl(oldUrl: String, newUrl: String) {
         database.writeTransaction(MEDIA_OBJECT_STORE_NAME) {
             val store = objectStore(MEDIA_OBJECT_STORE_NAME)
-            val value = store.get(Key(oldUrl))
-            if ((value == null).not()) { // because value != null is true for undefined
-                store.put(value, Key(newUrl))
-                store.delete(Key(oldUrl))
+            val value = store.get<Blob>(IDBValidKey(oldUrl))
+            if (value != null) {
+                store.put(value, IDBValidKey(newUrl))
+                store.delete(IDBValidKey(oldUrl))
             }
         }
     }
@@ -160,9 +174,9 @@ internal class IndexedDBMediaStore(
         val key = Random.nextString(12)
         val blob = database.writeTransaction(TMP_MEDIA_OBJECT_STORE_NAME) {
             val store = objectStore(TMP_MEDIA_OBJECT_STORE_NAME)
-            store.put(file, Key(key))
-            store.get(Key(key))
-        }.unsafeCast<Blob>()
+            store.put(file, IDBValidKey(key))
+            checkNotNull(store.get<Blob>(IDBValidKey(key)))
+        }
         return IndexeddbPlatformMediaTemporaryFileImpl(blob, key)
     }
 
@@ -174,7 +188,7 @@ internal class IndexedDBMediaStore(
             try {
                 database.writeTransaction(TMP_MEDIA_OBJECT_STORE_NAME) {
                     val store = objectStore(TMP_MEDIA_OBJECT_STORE_NAME)
-                    store.delete(Key(key))
+                    store.delete(IDBValidKey(key))
                 }
             } catch (_: Exception) {