+++ trixnity-idb-utils/src/commonMain/kotlin/net/folivo/trixnity/idb/utils/WrappedObjectStore.kt
@@ -1,0 +1,7 @@
+package net.folivo.trixnity.idb.utils
+
+import web.idb.IDBObjectStore
+
+value class WrappedObjectStore(val store: IDBObjectStore) {
+    companion object
+}